# specs/001-agentic-url-shortener/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic SDLC System - URL Shortener API
  version: 1.2.0
  description: |
    Contract-first API for the governed agentic SDLC orchestration system (control plane) and the
    URL shortener it governs (application plane). Feature `001-agentic-url-shortener`.

    **Versioning** (see `CHANGELOG.md` next to this file): URL major version (`/api/v1`) plus
    semantic `info.version`. Adding optional fields, endpoints, or error codes is MINOR; removing or
    renaming anything, or making an optional field required, is MAJOR and needs a change-control
    approval.

    **Errors** use RFC 9457 problem details (`application/problem+json`) with a stable machine-
    readable `code` and the request `correlationId`. Internal details are never exposed.

    **Authentication**: bearer tokens mapped to roles (`API_CONSUMER`, `REQUESTER`, `APPROVER`,
    `RELEASE_OWNER`, `AUDITOR`). Redirects (`GET /{code}`) and health/info are public.
  license:
    name: Proprietary assessment submission
servers:
  - url: http://localhost:8080
    description: Local developer machine
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workflows/{runId}:
    get:
      tags: [workflows]
      operationId: getRun
      summary: Inspect a run (status, stages, pending human actions)
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRun'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/workflows/{runId}/pause:
    post:
      tags: [operations]
      operationId: pauseRun
      summary: Pause a run (no new stages start; in-flight stages finish)
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorActionRequest'
      responses:
        '200':
          description: Run paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRun'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workflows/{runId}/resume:
    post:
      tags: [operations]
      operationId: resumeRun
      summary: Resume a paused run
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorActionRequest'
      responses:
        '200':
          description: Run resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRun'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workflows/{runId}/safe-stop:
    post:
      tags: [operations]
      operationId: safeStopRun
      summary: Safe-stop a run (halt, compensate, record terminal outcome)
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorActionRequest'
      responses:
        '200':
          description: Safe-stop initiated; the run reaches `SAFE_STOPPED`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRun'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/workflows/{runId}/audit:
    get:
      tags: [evidence]
      operationId: getAuditTrail
      summary: Append-only, hash-chained audit trail of a run
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Audit events ordered by sequence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/workflows/{runId}/audit/verification:
    get:
      tags: [evidence]
      operationId: verifyAuditTrail
      summary: Recompute the hash chain and report integrity
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Verification result (valid or first broken sequence)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerification'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/workflows/{runId}/summary:
    get:
      tags: [evidence]
      operationId: getFinalSummary
      summary: Final engineering summary of a terminal run (Markdown)
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Markdown summary
          content:
            text/markdown:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/reliability/report:
    get:
      tags: [evidence]
      operationId: getReliabilityReport
      summary: Reliability metrics computed from recorded evidence (demonstration data)
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReliabilityReport'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/policies:
    get:
      tags: [evidence]
      operationId: getPolicySet
      summary: Active versioned policy set
      responses:
        '200':
          description: Policy set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicySet'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/capabilities:
    get:
      tags: [evidence]
      operationId: listCapabilities
      summary: Capability release states of the application plane
      responses:
        '200':
          description: Capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capability'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque bearer token mapped to a principal and roles (demo tokens are labeled non-production).
...
  parameters:
...
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted (role or separation of duties)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: State conflict (alias in use, idempotency key reused, illegal state, deadline passed, release not ready)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457 problem detail. `code` values: VALIDATION_FAILED, URL_INVALID,
        URL_SCHEME_NOT_ALLOWED, URL_CREDENTIALS_NOT_ALLOWED, URL_HOST_MISSING, URL_HOST_NOT_ALLOWED,
        URL_TOO_LONG, INVALID_EXPIRY, INVALID_ALIAS, RESERVED_ALIAS, ALIAS_CONFLICT,
        INVALID_CLICK_LIMIT, CAPABILITY_NOT_AVAILABLE, INVALID_IDEMPOTENCY_KEY,
        IDEMPOTENCY_KEY_REUSED, RATE_LIMITED, CODE_GENERATION_EXHAUSTED, LINK_NOT_FOUND,
        LINK_EXPIRED, STORE_UNAVAILABLE, UNAUTHENTICATED, FORBIDDEN, SEPARATION_OF_DUTIES,
        RUN_NOT_FOUND, ARTIFACT_NOT_FOUND, CHANGE_REQUEST_NOT_FOUND, EXCEPTION_NOT_FOUND,
        RESOURCE_NOT_FOUND, METHOD_NOT_ALLOWED, UNSUPPORTED_MEDIA_TYPE,
        ILLEGAL_STATE, DEADLINE_PASSED, RELEASE_NOT_READY, RUN_TERMINAL, CONCURRENT_DECISION,
        FAULT_INJECTION_DISABLED, INTERNAL_ERROR.
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        correlationId: { type: string }
        retryAfterSeconds: { type: integer }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
...
    StageType:
      type: string
      enum:
        - REQUIREMENT_INGESTION
        - REQUIREMENT_ANALYSIS
        - CLARIFICATION
        - DECOMPOSITION
        - IMPACT_ANALYSIS
        - THREAT_ASSESSMENT
        - DESIGN
        - ARCHITECTURE_APPROVAL
        - CHANGE_APPROVAL
        - IMPLEMENTATION
        - TESTING
        - REGRESSION_TESTING
        - SECURITY_VERIFICATION
        - DOCUMENTATION
        - VALIDATION
        - COMPLIANCE_EVALUATION
        - RELEASE_APPROVAL
        - RELEASE
        - FINAL_SUMMARY
    RunStatus:
      type: string
      enum: [CREATED, RUNNING, AWAITING_HUMAN, PAUSED, COMPENSATING, COMPLETED, REJECTED, SAFE_STOPPED]
    StageStatus:
      type: string
      enum: [PENDING, READY, RUNNING, RETRY_WAIT, AWAITING_DECISION, SUCCEEDED, SKIPPED, FAILED, COMPENSATED, CANCELLED, REMOVED]
    Readiness:
      type: string
      enum: [READY, READY_WITH_ACCEPTED_LIMITATIONS, NOT_READY]
...
    WorkflowRun:
      type: object
      required: [runId, title, status, requestedBy, classification, planVersion, requirementVersion, policySetVersion, createdAt, stages, pendingActions]
      properties:
        runId: { type: string, format: uuid }
        requirementRef: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/RunStatus' }
        classification: { type: string, enum: [NEW_CAPABILITY, CHANGE_TO_EXISTING, UNDETERMINED] }
        requestedBy: { type: string }
        planVersion: { type: integer }
        requirementVersion: { type: integer }
        policySetVersion: { type: string }
        readiness: { $ref: '#/components/schemas/Readiness' }
        terminalOutcome: { type: string, enum: [COMPLETED, REJECTED, SAFE_STOPPED] }
        terminalReason: { type: string }
        manualInterventionRequired: { type: boolean }
        simulated: { type: boolean, description: True when the run carries injected faults }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        stages:
          type: array
          items:
            $ref: '#/components/schemas/Stage'
        pendingActions:
          type: array
          items:
            $ref: '#/components/schemas/PendingAction'
    Stage:
      type: object
      required: [key, type, status, dependsOn, generation, attempts, reused, degraded]
      properties:
        key: { $ref: '#/components/schemas/StageType' }
        type: { $ref: '#/components/schemas/StageType' }
        status: { $ref: '#/components/schemas/StageStatus' }
        awaiting: { type: string, enum: [APPROVAL, CLARIFICATION, POLICY_EXCEPTION, CHANGE_APPROVAL] }
        dependsOn:
          type: array
          items: { $ref: '#/components/schemas/StageType' }
        generation: { type: integer }
        attempts: { type: integer }
        reused: { type: boolean }
        degraded: { type: boolean }
        skipReason: { type: string }
        decisionDeadline: { type: string, format: date-time }
        lastFailureClass: { type: string, enum: [TRANSIENT, PERMANENT] }
        lastFailureReason: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    PendingAction:
      type: object
      required: [stageKey, actionType, requiredRole]
      properties:
        stageKey: { $ref: '#/components/schemas/StageType' }
        actionType: { type: string, enum: [APPROVAL, CLARIFICATION, POLICY_EXCEPTION, CHANGE_APPROVAL] }
        requiredRole: { type: string, enum: [APPROVER, RELEASE_OWNER, REQUESTER] }
        deadline: { type: string, format: date-time }
        reviewArtifacts:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactRef'
        referenceId: { type: string, description: Change request or policy exception id when applicable }
    ArtifactRef:
      type: object
      required: [artifactId, type, version, fingerprint]
      properties:
        artifactId: { type: string, format: uuid }
        type: { type: string }
        version: { type: integer }
        fingerprint: { type: string }
...
    PolicySet:
      type: object
      required: [version, policies]
      properties:
        version: { type: string }
        policies:
          type: array
          items:
            type: object
            required: [id, title, domain, severity]
            properties:
              id: { type: string }
              title: { type: string }
              domain: { type: string, enum: [SECURITY, PRIVACY, COMPLIANCE, LICENSING, TESTING, DOCUMENTATION, CHANGE_CONTROL, RELEASE] }
              severity: { type: string, enum: [MANDATORY, ADVISORY] }
              description: { type: string }
    OperatorActionRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 1000 }
...
    AuditEvent:
      type: object
      required: [chainId, seq, occurredAt, actorType, actorId, action, result, prevHash, hash]
      properties:
        chainId: { type: string }
        seq: { type: integer, format: int64 }
        occurredAt: { type: string, format: date-time }
        actorType: { type: string, enum: [HUMAN, AGENT, SYSTEM] }
        actorId: { type: string }
        action: { type: string }
        target: { type: string }
        fromState: { type: string }
        toState: { type: string }
        result: { type: string }
        reason: { type: string }
        details: { type: string }
        prevHash: { type: string }
        hash: { type: string }
    AuditVerification:
      type: object
      required: [chainId, valid, eventsChecked, verifiedAt, message]
      properties:
        chainId: { type: string }
        valid: { type: boolean }
        eventsChecked: { type: integer }
        firstBrokenSeq: { type: integer, format: int64 }
        verifiedAt: { type: string, format: date-time }
        message: { type: string }
    Capability:
      type: object
      required: [capabilityId, released]
      properties:
        capabilityId: { type: string }
        released: { type: boolean }
        parameters:
          type: object
          additionalProperties: true
        changedBy: { type: string }
        changedByRun: { type: string, format: uuid }
        changedAt: { type: string, format: date-time }
        reason: { type: string }
    ReliabilityReport:
      type: object
      required: [generatedAt, label, population, runs, stages, compensations, mttr, unrecoveredFailures, exclusions, latency, limitations]
      properties:
        generatedAt: { type: string, format: date-time }
        label: { type: string, example: 'DEMONSTRATION DATA - not production statistics' }
        population:
          type: object
          required: [runsConsidered, terminalRuns, activeRuns]
          properties:
            runsConsidered: { type: integer }
            terminalRuns: { type: integer }
            activeRuns: { type: integer }
            simulatedRuns: { type: integer }
        runs:
          type: object
          required: [completed, rejected, safeStopped, successRate, failureRate]
          properties:
            completed: { type: integer }
            rejected: { type: integer }
            safeStopped: { type: integer }
            successRate: { type: number }
            failureRate: { type: number }
        stages:
          type: object
          required: [attempts, retries, retryFrequency, fallbacks]
          properties:
            attempts: { type: integer }
            retries: { type: integer }
            retryFrequency: { type: number }
            fallbacks: { type: integer }
        compensations:
          type: object
          required: [runsWithCompensation, compensationFrequency, rollbacks]
          properties:
            runsWithCompensation: { type: integer }
            compensationFrequency: { type: number }
            rollbacks: { type: integer }
        mttr:
          type: object
          required: [recoveredEvents, totalRecoveryMillis, recoveryDurationsMillis, formula]
          properties:
            recoveredEvents: { type: integer }
            totalRecoveryMillis: { type: integer, format: int64 }
            mttrMillis: { type: number }
            recoveryDurationsMillis:
              type: array
              items: { type: integer, format: int64 }
            byMechanism:
              type: object
              additionalProperties: { type: integer }
            formula: { type: string }
        unrecoveredFailures:
          type: object
          required: [count, events]
          properties:
            count: { type: integer }
            events:
              type: array
              items:
                type: object
                properties:
                  runId: { type: string, format: uuid }
                  stageKey: { type: string }
                  cause: { type: string }
                  simulated: { type: boolean }
                  detectedAt: { type: string, format: date-time }
        exclusions:
          type: array
          items:
            type: object
            required: [reason, count]
            properties:
              reason: { type: string }
              count: { type: integer }
        latency:
          type: object
          required: [sampleSize, excludesHumanWait]
          properties:
            sampleSize: { type: integer }
            minMillis: { type: integer, format: int64 }
            p50Millis: { type: integer, format: int64 }
            p95Millis: { type: integer, format: int64 }
            maxMillis: { type: integer, format: int64 }
            excludesHumanWait: { type: boolean }
        limitations:
          type: array
          items: { type: string }